requirements.txt
openai
gradio
python-dotenv
chromadb
numpy
litellm
pydantic
tenacity
tqdm